temporarly remove modifications for empact-css
extra css (like for empcat) can be done by adding new files and just
edit the one bootstrap_and_overrides.css.scss

all other changes for empact moved to 'Settings' and are programmatically
in this version.

this now get's merged into master and from now on we can sync
'empact' with master if new features are added to master.